Gradle structure
Reformat code

Gradle file update

Cleaned out old trash

Correct package name

Read test resources recursive
