+++ Test Results - All_in_ImageRenamer.html
@@ -640,10 +640,10 @@ jQuery.cookie = function(name, value, options) {
 <br/>	at org.junit.runner.JUnitCore.run(JUnitCore.java:157)
 <br/>	at com.intellij.junit4.JUnit4IdeaTestRunner.startRunnerWithArgs(JUnit4IdeaTestRunner.java:77)
 <br/>	at com.intellij.rt.execution.junit.JUnitStarter.prepareStreamsAndStart(JUnitStarter.java:195)
-<br/>	at com.intellij.rt.execution.junit.JUnitStarter.se.kotlinski.imageRenamer.main(JUnitStarter.java:63)
+<br/>	at com.intellij.rt.execution.junit.JUnitStarter.main.imageRenamer.main(JUnitStarter.java:63)
 <br/>	at sun.reflect.NativeMethodAccessorImpl.invoke0(Native Method)
 <br/>	at sun.reflect.NativeMethodAccessorImpl.invoke(NativeMethodAccessorImpl.java:39)
-<br/>	at com.intellij.rt.execution.application.AppMain.se.kotlinski.imageRenamer.main(AppMain.java:120)
+<br/>	at com.intellij.rt.execution.application.AppMain.main.imageRenamer.main(AppMain.java:120)
 <br/></span>
 </li>
 </ul>
