+++ build.gradle
@@ -3,20 +3,30 @@ repositories {
 }
 
 apply plugin: 'java'
+apply plugin: 'application'
 
-task wrapper(type: Wrapper) {
-    gradleVersion = '1.10'
+mainClassName = "se.kotlinski.imageRenamer.MainRenamer"
+
+jar {
+    manifest.attributes provider: 'gradle'
 }
-
-
 
 sourceSets {
     main {
-        java.srcDir file('src/')
-
+        java {
+            srcDir 'src/java'
+        }
+        resources {
+            srcDir 'src/resources'
+        }
     }
     test {
-        java.srcDir file('test/')
+        java {
+            srcDir 'test/java'
+        }
+        resources {
+            srcDir 'src/resources'
+        }
     }
 }
