--- settings.json
@@ -1,1 +1,0 @@
-{"samsungPath":"aaa","dropboxPath":"bbb","outputPath":"ccc"}
