+++ src/main/java/se/kotlinski/imageRenamer/utils/Constants.java
@@ -1,0 +1,15 @@
+package se.kotlinski.imageRenamer.utils;
+
+/**
+ * Created with IntelliJ IDEA.
+ * User: Simon
+ * Date: 2013-12-23
+ * Time: 23:30
+ * To change this template use File | MainRenamer.imageRenamer.Settings | File Templates.
+ */
+public class Constants {
+	public static final String USER_FOLDER = "user.folder";
+	public static final String PATH_INPUT_TEST = System.getProperty("user.dir") + "//src//test//resources//inputImages";
+
+
+}
