+++ src/test/resources/inputImages/structure/2013-10-03 13.43.20.jpg
