+++ src/test/resources/inputImages/structure/2013-10-26 20.20.46.jpg
