+++ src/test/resources/inputImages/structure2/2014-02-22 11.48.47-2.jpg
