+++ src/test/resources/inputImages/structure2/structure2.1/2014-02-22 11.48.48-1.jpg
