+++ src/test/resources/inputImages/structure2/structure2.2/2014-03-02 01.09.34.jpg
